Move lost errors in with the rest
