+++ src/transport/errors/index.ts
@@ -1,3 +1,5 @@
+export * from "./AccountableError.js";
 export * from "./NetworkError.js";
 export * from "./NotImplementedError.js";
 export * from "./UnexpectedResponseError.js";
+export * from "./UnreachableError.js";
