+++ src/transport/storage.ts
@@ -1,5 +1,6 @@
 import type { AccountableDB } from "./db";
-import { AccountableError, previousStats } from "./db";
+import { AccountableError } from "./errors/index.js";
+import { previousStats } from "./db";
 import { deleteAt, downloadFrom, uploadTo } from "./networking";
 
 /**
